If npm install fails, use "npm run reinstall" to ensure we start at clean slate
This addresses issue #158, but only when you are in trouble.

This approach has a tradeoff: Firstly it is not done automatically, which requires the use to think after being in trouble

On the other hand, this approach saves time since a full clearing of caches and installation of all modules takes time.

Furthermore, adding the command as "npm run reinstall" makes it explicit what should and could be done