Faster loading of tasks
(skip unused)

This has annoyed me forever. On my machine, calling the custom "task()" method takes roughly half a
second.

Depending on which gulp task you execute, there is no reason to load these tasks in the top of
the gulpfile. I wrap them in a function, which hides the underlying true problem.

The real problem is of course that the loading of tasks take a long time, and hides unused tasks so that we never delete them.

But that will be a separate problem requiring a bigger fix.