Fix docker build (#2325)
Despite the fact that the package 'browserslist' is a transitive dependency of:
- autoprefixer@9.0.1
- cssnano@4.0.3

It needs to be added manually to our package.json to be installed in the docker environment. This issue doesn't seem related to 'optionalDependencies' since the chain of dependencies from cssnano is compulsory all the way (not even a single 'devDependency').